requirements.txt
python-dotenv
aiogram
pillow
